[EI-161] [web] feat: Add SourceIndicatorSearchListItem component
