+++ web/app/ui/components/numerical-guidance/molecule/source-indicator-search-list.tsx
@@ -4,46 +4,16 @@ import { ListChildComponentProps } from 'react-window';
 import { IndicatorInfoResponse } from '@/app/store/querys/numerical-guidance/indicator.query';
 import TinyInput from '../../view/atom/tiny-input/tiny-input';
 import { SearchIcon } from '@heroicons/react/solid';
-import SelectableItem from '../../view/atom/selectable-item';
 import { useSelectedCustomForecastIndicatorViewModel } from '@/app/business/hooks/custom-forecast-indicator/use-selected-custom-forecast-indicator-view-model';
-import ToolTip from '../../view/atom/tooltip';
-import { SparklesIcon } from '@heroicons/react/outline';
-import Icon from '../../view/atom/icons/variant-icon';
+import SourceIndicatorSearchListItem from '../atom/source-indicator-search-list-item';
 
 export default function SourceIndicatorSearchList() {
-  const { selectedCustomForecastIndicator } = useSelectedCustomForecastIndicatorViewModel();
   const { indicatorList } = useIndicatorList();
 
   const render = ({ index, style, data }: ListChildComponentProps<IndicatorInfoResponse[]>) => {
     const indicator = data[index];
 
-    const handleSelect = () => {
-      // logic: 재료 지표 선택
-      console.log('handleSelect');
-    };
-
-    const handleDeSelect = () => {
-      // logic: 재료 지표 선택 해제
-      console.log('handleDeSelect');
-    };
-
-    const indicatorDisplayName = `${indicator.ticker}(${indicator.name})`;
-
-    const isTargetIndicator = selectedCustomForecastIndicator?.targetIndicatorId === indicator.id;
-    const isSelected = selectedCustomForecastIndicator?.sourceIndicatorIds?.includes(indicator.id) || isTargetIndicator;
-
-    return (
-      <SelectableItem selected={isSelected} onSelect={handleSelect} onDeSelect={handleDeSelect} style={style}>
-        <div className="flex w-full items-center justify-between">
-          <span className="text-xs">{indicatorDisplayName}</span>
-          {isTargetIndicator ? (
-            <ToolTip message="대상 지표입니다.">
-              <Icon size={'xs'} icon={SparklesIcon} />
-            </ToolTip>
-          ) : null}
-        </div>
-      </SelectableItem>
-    );
+    return <SourceIndicatorSearchListItem item={indicator} style={style} />;
   };
 
   return (
